facefind_back/requirements.txt
flask
flask-cors
fastapi 
uvicorn 
supabase
python-dotenv
bcrypt
opencv-python
numpy
face-recognition
requests